Create separate connection to consumer Group Coordinator
Since JoinGroupRequests may block for up to max.poll.interval.ms,
which may be set very high (hours..), any sub-sequent requests
on the same connection, such as Metadata refreshes, would time out
and tear down the connection, triggering another rebalance.

This commit adds support for logical brokers, in this case
a GroupCoordinator broker, which initially doesn't have an address
but is updated by the cgrp to connect to the group coordinator.
This connection is then used for all group-related protocol requests,
Join, Sync, Leave, OffsetCommit, leaving the standard broker
connections for other protocol use.

As a side-effect this commit also deletes legacy cgrp code from when
it used to run on the broker thread.